First attempt at templates
The current issue is that the generated strings don't work with the formatting of the variables like `%d` and `%s`. I'm not currently sure how to resolve that and generate support for those dynamically, so I'll try to come back and finish this when I have time in the future.